dugit/isomorphic (response processor) wrapped inside git_utilis
